requirements.txt
google-api-core==2.12.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-storage==2.13.0
numpy==1.26.1
opencv-python==4.8.1.78
pandas==2.1.1
